requirement.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
Django==1.9.7
django-autoslug==1.9.3
django-crispy-forms==1.6.0
django-taggit==0.20.0
django-taggit-autosuggest==0.3.0
django-tinymce==2.3.0
et-xmlfile==1.0.1
ipython==4.2.0
ipython-genutils==0.1.0
jdcal==1.2
openpyxl==2.3.5
pathlib2==2.1.0
pexpect==4.1.0
pickleshare==0.7.2
Pillow==3.2.0
psycopg2==2.6.1
ptyprocess==0.5.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
xlwt==1.1.2
